Rename plural components/containers with -List
