--- src/routes/NPCs/routes/Profile/components/Profile/index.scss
@@ -1,7 +1,0 @@
-@import 'base';
-
-
-.empty {
-  @extend .flex-row-center;
-  padding-top: 8rem;
-}
