--- src/routes/NPCs/routes/Profile/index.js
@@ -1,22 +1,0 @@
-export default store => ({
-  path: ':profilename',
-  /*  Async getComponent is only invoked when route matches   */
-  getComponent(nextState, cb) {
-    /*  Webpack - use 'require.ensure' to create a split point
-        and embed an async module loader (jsonp) when bundling   */
-    require.ensure(
-      [],
-      require => {
-        /*  Webpack - use require callback to define
-          dependencies for bundling   */
-        const Profile = require('containers/Profile').default
-
-        /*  Return getComponent   */
-        cb(null, Profile)
-
-        /* Webpack named bundle   */
-      },
-      'Profile'
-    )
-  }
-})
