+++ src/store/reducers/form.js
@@ -1,7 +1,6 @@
-import { combineReducers } from 'redux'
 import { reducer as formReducer } from 'redux-form'
 
-import { newProfileFormReducer as newProfile } from 'containers/Profiles/reducers'
+import { newProfileFormReducer as newProfile } from 'containers/ProfileList/reducers'
 
 const reducer = formReducer.plugin({
   newProfile
