+++ src/store/reducers/modal.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux'
 
-import { newProfileModalReducer as newProfile } from 'containers/Profiles/reducers'
+import { newProfileModalReducer as newProfile } from 'containers/ProfileList/reducers'
 
 const reducer = combineReducers({
   newProfile
